Refactor expense store into shared provider
