+++ src/App.tsx
@@ -12,6 +12,7 @@ import Login from "./pages/Login";
 import Register from "./pages/Register";
 import PrivateRoute from "./components/PrivateRoute";
 import { AuthProvider } from "@/hooks/useAuth";
+import { ExpenseProvider } from "@/hooks/useExpenseStore";
 import Layout from "@/components/Layout";
 
 const queryClient = new QueryClient();
@@ -19,26 +20,28 @@ const queryClient = new QueryClient();
 const App = () => (
   <QueryClientProvider client={queryClient}>
     <AuthProvider>
-      <TooltipProvider>
-        <Toaster />
-        <Sonner />
-        <BrowserRouter>
-          <Routes>
-            <Route path="/login" element={<Login />} />
-            <Route path="/register" element={<Register />} />
-            <Route element={<PrivateRoute />}>
-              <Route element={<Layout />}>
-                <Route index element={<Index />} />
-                <Route path="/projected" element={<ProjectedExpenses />} />
-                <Route path="/month/:year/:month" element={<MonthDetail />} />
-                <Route path="/category/:category" element={<CategoryDetail />} />
-                {/* ADD ALL CUSTOM ROUTES ABOVE THE CATCH-ALL "*" ROUTE */}
-                <Route path="*" element={<NotFound />} />
+      <ExpenseProvider>
+        <TooltipProvider>
+          <Toaster />
+          <Sonner />
+          <BrowserRouter>
+            <Routes>
+              <Route path="/login" element={<Login />} />
+              <Route path="/register" element={<Register />} />
+              <Route element={<PrivateRoute />}>
+                <Route element={<Layout />}>
+                  <Route index element={<Index />} />
+                  <Route path="/projected" element={<ProjectedExpenses />} />
+                  <Route path="/month/:year/:month" element={<MonthDetail />} />
+                  <Route path="/category/:category" element={<CategoryDetail />} />
+                  {/* ADD ALL CUSTOM ROUTES ABOVE THE CATCH-ALL "*" ROUTE */}
+                  <Route path="*" element={<NotFound />} />
+                </Route>
               </Route>
-            </Route>
-          </Routes>
-        </BrowserRouter>
-      </TooltipProvider>
+            </Routes>
+          </BrowserRouter>
+        </TooltipProvider>
+      </ExpenseProvider>
     </AuthProvider>
   </QueryClientProvider>
 );
